# dist/swagger_output.json
{
    "swagger": "2.0",
    "info": {
        "title": "Car Marketplace API",
        "description": "API documentation for Car Marketplace application",
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/api",
    "tags": [
        {
            "name": "Authentication",
            "description": "User authentication endpoints"
        },
        {
            "name": "Cars",
            "description": "Car management endpoints"
        }
    ],
    "schemes": [
        "http"
    ],
    "securityDefinitions": {
        "bearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "scheme": "bearer",
            "in": "header"
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/signup": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "example": "any"
                                },
                                "password": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "411": {
                        "description": "Length Required"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "example": "any"
                                },
                                "password": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/createcars": {
            "post": {
                "description": "",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/cars": {
            "get": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/cars/test-route": {
            "put": {
                "description": "",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/cars/{carId}": {
            "put": {
                "description": "",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "description": "",
                "parameters": [
                    {
                        "name": "carId",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/cars/search": {
            "get": {
                "description": "",
                "responses": {
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "definitions": {
        "Car": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "example": "object"
                },
                "properties": {
                    "type": "object",
                    "properties": {
                        "title": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "string"
                                }
                            }
                        },
                        "description": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "string"
                                }
                            }
                        },
                        "images": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "array"
                                },
                                "items": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "string"
                                        },
                                        "format": {
                                            "type": "string",
                                            "example": "binary"
                                        }
                                    }
                                }
                            }
                        },
                        "tags": {
                            "type": "object",
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "example": "object"
                                },
                                "properties": {
                                    "type": "object",
                                    "properties": {
                                        "carType": {
                                            "type": "object",
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "string"
                                                }
                                            }
                                        },
                                        "company": {
                                            "type": "object",
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "string"
                                                }
                                            }
                                        },
                                        "dealer": {
                                            "type": "object",
                                            "properties": {
                                                "type": {
                                                    "type": "string",
                                                    "example": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}